rename downloader mod and main
